s47-vibe-coding-livekit/agent/requirements.txt
livekit-agents>=1.5.0
livekit-plugins-silero>=1.5.0
livekit-plugins-openai>=1.5.0
livekit-plugins-deepgram>=1.5.0
httpx>=0.27
